Issue 38 (#40)
* [issue-38][ RJ ] seed file

* [issue-38][ RJ ] added report input type

* [issue-38][ RJ ] Updated resolver

The resolved needed to be changed due to the changed input type.

* [issue-38][ RJ ] updated input type

The update method needed to be refactored. I changed the name to call #update and not edit.

* [issue-38][ RJ ] User can flag a report

Ability to flagg a report. Cannot be unflagged. ... TODO, Ass ability to unflag/

* [issue-38][ RJ ] refactored

Moved theAR object to its own class. This will allow ability to expand and DRY up the error portion.

* [issue-38][ RJ ] rubocop

rubocop

* [issue-38][ RJ ] Work... lots of work.

A lot of refcatoring as well as the testing of the inputs.